# Make prefiltering faster