--- a/src/b-splines/constant.jl
+++ b/src/b-splines/constant.jl
@@ -1,28 +1,28 @@
 immutable Constant <: Degree{0} end
 
-function define_indices{GT<:GridType}(::Type{BSpline{Constant,None}}, ::Type{GT}, N)
+function define_indices(::Type{BSpline{Constant}}, N)
     :(@nexprs $N d->(ix_d = clamp(round(Int, real(x_d)), 1, size(itp, d))))
 end
 
-function coefficients{GT<:GridType}(::Type{BSpline{Constant,None}}, ::Type{GT}, N)
-    :(@nexprs $N d->($(coefficients(BSpline{Constant,None}, GT, N, :d))))
+function coefficients(::Type{BSpline{Constant}}, N)
+    :(@nexprs $N d->($(coefficients(BSpline{Constant}, N, :d))))
 end
 
-function coefficients{GT<:GridType}(::Type{BSpline{Constant,None}}, ::Type{GT}, N, d)
+function coefficients(::Type{BSpline{Constant}}, N, d)
     sym, symx = symbol(string("c_",d)), symbol(string("x_",d))
     :($sym = 1)
 end
 
-function gradient_coefficients{GT<:GridType}(::Type{BSpline{Constant,None}}, ::Type{GT}, N, d)
+function gradient_coefficients(::Type{BSpline{Constant}}, N, d)
     sym, symx = symbol(string("c_",d)), symbol(string("x_",d))
     :($sym = 0)
 end
 
-function index_gen{IT<:BSpline{Constant}}(::Type{IT}, N::Integer, offsets...)
+function index_gen(::Type{BSpline{Constant}}, N::Integer, offsets...)
     if (length(offsets) < N)
         d = length(offsets)+1
         sym = symbol("c_"*string(d))
-        return :($sym * $(index_gen(IT, N, offsets..., 0)))
+        return :($sym * $(index_gen(BSpline{Constant}, N, offsets..., 0)))
     else
         indices = [offsetsym(offsets[d], d) for d = 1:N]
         return :(itp.coefs[$(indices...)])