… src/b-splines/linear.jl
… src/b-splines/prefiltering.jl
… src/b-splines/quadratic.jl
… src/constant.jl
… src/extrapolation.jl
… src/extrapolation/constant.jl
… src/extrapolation/extrapolation.jl
… src/extrapolation/indexing.jl
… src/filter1d.jl
… test/b-splines/constant.jl
… test/b-splines/linear.jl
… test/b-splines/quadratic.jl
… test/b-splines/runtests.jl
… test/extrapolation/runtests.jl
… test/plots/b-splines/sanity.jl
… test/quadratic.jl
… test/runtests.jl