@timholy 2015-05-16
I should add that this makes the prefiltering in Interpolations faster than Grid, too. Grid was a bit faster than Interplations master, but not by much; in a 2d case, I got 0.19 s for master Interpolations, .17 s for Grid, and 0.12 for this PR.