@timholy 2015-05-16
Note mostly to self: Grid seems to be able to do without Woodbury in a few more cases than Interpolations. Try to figure out why (important since it has a big performance impact).