@timholy 2015-05-20
OK, this has been rebased on #31. I'll probably wait until that's merged to work on #34.