@timholy 2015-06-01
As an update, once https://github.com/JuliaLang/METADATA.jl/pull/2670 merges, I will be able to contribute a further improvement in the Woodbury prefiltering.